Rename common/ to primitives/
"common" is ambiguous about abstraction level. "primitives" is more
descriptive: the directory contains foundational data structures and
building blocks (data paths, errors, reactivity, cancellation) with
no business logic. The restrictive name also discourages dumping
higher-level utilities into it as the library grows.
