+++ packages/a2ui_core/lib/src/core/surface_model.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import '../common/event_notifier.dart';
+import '../primitives/event_notifier.dart';
 import 'catalog.dart';
 import 'common.dart';
 import 'component_model.dart';
