+++ packages/a2ui_core/lib/src/primitives/event_notifier.dart
