+++ packages/a2ui_core/lib/src/primitives/reactivity.dart
