+++ packages/a2ui_core/lib/src/processing/basic_functions.dart
@@ -3,10 +3,11 @@
 // found in the LICENSE file.
 
 import 'package:json_schema_builder/json_schema_builder.dart';
-import '../common/cancellation.dart';
-import '../common/reactivity.dart';
+
 import '../core/catalog.dart';
 import '../core/contexts.dart';
+import '../primitives/cancellation.dart';
+import '../primitives/reactivity.dart';
 import 'expressions.dart';
 
 class FormatStringFunction extends FunctionImplementation {
